% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicglmnet.R
\name{bdiag_m}
\alias{bdiag_m}
\title{bdiag_m}
\usage{
bdiag_m(lmat)
}
\arguments{
\item{lamt:}{A list of block matrices.}
}
\value{
a matrix
}
\description{
FUNCTION:Fast version of block diag. Copyright (C) 2016 Martin Maechler, ETH Zurich
}
