% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicglmnet.R
\name{bicglmnet}
\alias{bicglmnet}
\title{bicglmnet}
\usage{
bicglmnet(x, y, weight, indmisfun, bdiag_m)
}
\arguments{
\item{x:}{Predictor matrix.}

\item{y:}{Reponse.}

\item{indmisfunction:}{To indicate which value of one SNP belongs to.}

\item{bdiag_M:}{Fast version of block diag.}
}
\value{
a list contained the optimal ridge regression model
}
\description{
FUNCTION: Defined BIC for binomial distribution and trinomial distribution.
}
