% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varselrfmisMIfun.R
\name{detctnafun}
\alias{detctnafun}
\title{detctnafun}
\usage{
detctnafun(predt, reps)
}
\arguments{
\item{predt:}{dataframe of SNPs}

\item{resps:}{phenotype factor}
}
\value{
a list which contains "detna" indicating which individual have missing values and "misind" indicating which entry is missing in the dataset
}
\description{
FUNCTION: To generate a list of which includes a vector for indicating whether SNPs and phenotype are missing and a matrix for indicating what components in genomic matrix and phenotype are missing.
}
