% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varselrfmisMIfun.R
\name{dummytran}
\alias{dummytran}
\title{dummytran}
\usage{
dummytran(cl)
}
\arguments{
\item{cl:}{a vector of SNP}
}
\value{
a matrix which contains the transformed dummy data
}
\description{
FUNCTION: To convert each SNP vector into the dummy variable
}
