% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagename-data.R
\docType{data}
\name{gene_phe}
\alias{gene_phe}
\title{A list contains the simulated genotype data with missing values, phenotype data with missing values}
\format{
a list
}
\description{
A list contains the simulated genotype data with missing values, phenotype data with missing values
}
