% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensetf.R
\name{gensetf}
\alias{gensetf}
\title{gensetf}
\usage{
gensetf(gvec)
}
\arguments{
\item{gvec:}{gvec: vector consisting of numbers of levels of all factors.}
}
\value{
a matrix containing all possible combination
}
\description{
Function name: genset.f. It is used to generate all possible conbinations of factor levels in an experiment.Written by Guoqi Qian, 17/11/2002.
}
