% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varselrfmisMIfun.R
\name{jimpmisfun}
\alias{jimpmisfun}
\title{jimpmisfun}
\usage{
jimpmisfun(
  misindcom,
  xycat,
  dectnalist,
  ncol.predyimp,
  predymisind,
  complet.ind,
  ridgimp,
  ridgselectedvars,
  optmethod,
  gensetf,
  dummytran,
  indmisfun,
  mulimpmisfun
)
}
\arguments{
\item{misindcom:}{A missing indicator matrix for a dataframe of SNPs and phenotype after multiple imputation.}

\item{xycat:}{a vector indicating the number of levels in each SNPs and phenotype.}

\item{dectnalist:}{A vector which indicates whether SNPs and phenotype are missing.}

\item{ncol.predyimp:}{The number of SNPs and phenotype.}

\item{predymisind:}{The dataframe contains SNP data, phenotype and missing indicator matrix which only contains SNPs and phenotype with missing values.}

\item{complet.ind:}{A vector which indicates whether each individual contain missing values.}

\item{ridgimp:}{A list of ridge models with the optimal tunning parameter.}

\item{ridgselectedvars:}{A list which contains selected important variables in a series of models for missing SNPs, corresponding missing indicators and phenotype.}

\item{gensetf:}{A function for generating all possible conbinations of factor levels in an experiment}

\item{dummytran:}{A function which converts each SNP vector into the dummy variable.}

\item{indmisfun:}{A function which indicates the value of a SNP ("0", "1" or "2").}

\item{mulimpmisfun:}{A function which generates multiple imputation dataset in each individual with missing values.}
}
\value{
a dataframe named "mdatamis" comprised of multiple imputed data and the observed data with the corresponding weights
}
\description{
FUNCTION: To multiple impute missing values in SNPs and phenotype.
}
