% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varselrfmisMIfun.R
\name{mulimpmisfun}
\alias{mulimpmisfun}
\title{mulimpmisfun}
\usage{
mulimpmisfun(
  submisrow,
  xycat,
  dectnalist,
  ncol.predyimp,
  xyindrow,
  ridgimp,
  ridgselectedvars,
  optmethod,
  gensetf,
  dummytran,
  indmisfun
)
}
\arguments{
\item{submisrow:}{One row of "misindcom" matrix.}

\item{xycat:}{a vector indicating the number of levels in each SNPs and phenotype.}

\item{dectnalist:}{A vector which indicates whether SNPs and phenotype are missing.}

\item{ncol.predyimp:}{The number of SNPs and phenotype.}

\item{xyindrow:}{One row of multiple imputed dataframe of SNPs and phenotype.}

\item{ridgimp:}{A list of ridge models with the optimal tunning parameter.}

\item{ridgselectedvars:}{A list which contains selected important variables in a series of models for missing SNPs, corresponding missing indicators and phenotype.}

\item{gensetf:}{A function for generating all possible conbinations of factor levels in an experiment.}

\item{dummytran:}{A function which converts each SNP vector into the dummy variable.}

\item{indmisfun:}{A function which indicates the value of a SNP ("0", "1" or "2").}
}
\value{
a dataframe named "mulimp" which contains the multiple imputed data with the corresponding weights
}
\description{
FUNCTION: To generate multiple imputation dataset in each individual with missing values.
}
