% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varselrfmisMIfun.R
\name{priowinitfun}
\alias{priowinitfun}
\title{priowinitfun}
\usage{
priowinitfun(predt, resps, misind, complet.ind)
}
\arguments{
\item{predt:}{dataframe of SNPs.}

\item{resps:}{phenotype factor.}

\item{misind:}{a list of length 2. The first part is a matrix which is composed of "1" and "0" indicating whether each component is missing. The second part is a vector of indicating which component in the phenotype is missing.}

\item{complet.ind:}{a vector which indicates whether each individual contain missing values.}
}
\value{
a list containing "predimp", "predy", "misidimp" which would be fitted by the models for missingness mechanism and "weight".
}
\description{
FUNCTION: To initially add some weights to individuals which contain few missing values.
}
