% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varselrfmisMIfun.R
\name{rfmisinitfun}
\alias{rfmisinitfun}
\title{rfmisinitfun}
\usage{
rfmisinitfun(predt, resps, misind, detna)
}
\arguments{
\item{predt:}{dataframe of SNPs.}

\item{resps:}{phenotype factor.}

\item{misind:}{a list of length 2. The first part is a matrix which is composed of "1" and "0" indicating whether each component is missing. The second part is a vector of indicating which component in the phenotype is missing.}

\item{detna:}{a list of length 2. The first vector indicates which SNP contain missing values, and the second vector indicates whether phenotype is missing.}
}
\value{
a list which contains initialized "predimp" and "predy"
}
\description{
FUNCTION: To initially impute missing values which are present in the genomic data and phenotype.
}
