% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varselrfmisMIfun.R
\name{varselrfmisfun}
\alias{varselrfmisfun}
\title{varselrfmisfun}
\usage{
varselrfmisfun(
  predt,
  resps,
  misind,
  detna,
  iter.varsel = 10,
  initmethod = "priorweight",
  optmethod = "bic",
  dummytran,
  vsrfmisfun,
  jimpmisfun,
  mulimpmisfun,
  imputerfmisfun,
  varSelRF_mis,
  gensetf,
  bicglmnet,
  bdiag_m
)
}
\arguments{
\item{predt:}{dataframe of SNPs.}

\item{resps:}{phenotype factor.}

\item{misind:}{A list of length 2. The first part is a matrix which is composed of "1" and "0" indicating whether each component is missing. The second part is a vector of indicating which component in the phenotype is missing.}

\item{detna:}{A list of length 2. The first vector indicates which SNP contain missing values, and the second vector indicates whether phenotype is missing.}

\item{iter.varsel:}{The number of iterations.}

\item{initmethod:}{"priorweight" or "init". "priorweight" is used for initializing SNPs with few missing values. If there are no SNPs with few missing values, choose "init".}

\item{optmethod:}{"bic" or "cv". It is used for choosing the optimal tunning parameter.}

\item{dummytran:}{A function which converts each SNP vector into the dummy variable.}

\item{vsrfmisfun:}{A function which selects important SNPs using the random forests approach introduced by Diaz}

\item{jimpmisfun:}{A function to multiple impute missing values in SNPs and phenotype.}

\item{mulimpmisfun:}{A function which generates multiple imputation dataset in each individual with missing values.}

\item{imputerfmisfun:}{A function to choose the optimal tunning parameter.}

\item{varSelRF_mis:}{A function for selecting important variables.}

\item{gensetf:}{A function for generating all possible conbinations of factor levels in an experiment.}

\item{bicglmnet:}{A function which chooses an optimal tunning parameter using the model criterion BIC.}

\item{bdiag_m:}{A function which generates a block diagonal matrix.}
}
\value{
a list containing the final selected result
}
\description{
FUNCTION: To select important SNPs with non-ignorable missing values using random forests.
}
