% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varselrfmisMIfun.R
\name{vsrfmisfun}
\alias{vsrfmisfun}
\title{vsrfmisfun}
\usage{
vsrfmisfun(predyimp, predymisind, dectnalist, na.phe, weight, varSelRF_mis)
}
\arguments{
\item{predyimp:}{The dataframe contains SNP and phenotype data after multiple imputation.}

\item{predymisind:}{The dataframe contains SNP data, phenotype and missing indicator matrix which only contains SNPs and phenotype with missing values.}

\item{dectnalist:}{A vector which indicates whether SNPs and phenotype are missing.}

\item{na.phe:}{A missing indicator of phenotype.}

\item{weight:}{A vector indicating each weight in every individual.}

\item{varSelRF_mis:}{A function for selecting important variables.}
}
\value{
a list containing the results of selected variables using random forest
}
\description{
FUNCTION: To select important SNPs using the random forests approach introduced by Diaz
}
